requirements.txt
Flask>=2.0
google-generativeai>=0.6.0
SpeechRecognition>=3.8.1
gTTS>=2.3.0
pydub>=0.25.1
gunicorn